% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_subset.R
\name{FlowSOM_subset}
\alias{FlowSOM_subset}
\title{Take a subset from the flowSet}
\usage{
FlowSOM_subset(fsom, file)
}
\arguments{
\item{fsom}{FlowSOM object as generated by the FlowSOM function}

\item{file}{File to select}
}
\description{
Take a subset from the flowSet
}
