% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating_subset.R
\name{gating_subset}
\alias{gating_subset}
\title{Select only a subset of cells}
\usage{
gating_subset(flowjo_res, gate)
}
\arguments{
\item{flowjo_res}{The parsed flowjo result, as generated by parse_flowjo}

\item{gate}{The name of the gate that should be used for selection}
}
\value{
Returns a list with two elements: flowFrame/flowSet and gates. Both are
subsetted to contain only the cells of interest.
}
\description{
Given a parsed flowjo result,
return only those cells which fall into a specific gate
}
