% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Get counts per input file}
\usage{
get_counts(fsom)
}
\arguments{
\item{fsom}{FlowSOM object as returned by the FlowSOM() function}
}
\value{
A list with cluster counts and metacluster counts
}
\description{
Computes counts per file per cluster and metacluster
}
