% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pctgs.R
\name{get_pctgs}
\alias{get_pctgs}
\title{Get percentages per input file}
\usage{
get_pctgs(fsom, meta_names = NULL)
}
\arguments{
\item{fsom}{FlowSOM object as returned by the FlowSOM() function}

\item{meta_names}{Names to assign to the metaclusters. If NULL, numbers
are used.}
}
\value{
A list with cluster percentages and metacluster percentages
}
\description{
Computes percentages per file per cluster and metacluster
}
