% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_flowjo.R
\name{parse_flowjo}
\alias{parse_flowjo}
\title{Parse a FlowJo workspace for one specific file}
\usage{
parse_flowjo(files, wsp_file, group = "All Samples", plot = FALSE)
}
\arguments{
\item{wsp_file}{Path to the FlowJo workspace}

\item{group}{Group the file belongs to in the FlowJo workspace.
Default is "All Samples"}

\item{file}{Path to the fcs file of interest}
}
\value{
Returns a list with two elements: flowFrame and gates.
\code{flowFrame} contains the compensated and transformed fcs file, as set
up in the FlowJo workspace.
\code{gates} is a matrix in which each row corresponds with an individual
cell and each column with an individual gate, containing TRUE/FALSE values
indicating whether the cell falls into that gate
}
\description{
Extract the compensated and transformed data and all gate labels.
}
