% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate.R
\name{plot_aggregate}
\alias{plot_aggregate}
\title{Plot Aggregate}
\usage{
plot_aggregate(flowset, labels = NULL, output_image = "aggregate.png")
}
\arguments{
\item{flowset}{FlowSet to plot}

\item{labels}{If not NULL, colors the files per label.}

\item{output_image}{Path to write the png figure to.}
}
\value{
nothing but a plot is generated
}
\description{
Plot aggregate figure
}
