% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNE_manual_gating.R
\name{plotSNE_manual_gating}
\alias{plotSNE_manual_gating}
\title{plotSNE_manual_gating}
\usage{
plotSNE_manual_gating(tsne_result, cell_ids)
}
\arguments{
\item{tsne_result}{The resulting object of the Rtsne::Rtsne function}

\item{cell_ids}{The labels of the cells, as assigned by manual gating}
}
\value{
ggplot of the tSNE result colored according to the manual gating.
}
\description{
Function to plot the result of a tSNE, colored according to the manual gating.
}
\examples{


}
