% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNE_marker.R
\name{plotSNE_marker}
\alias{plotSNE_marker}
\title{plotSNE_marker}
\usage{
plotSNE_marker(flowframe, tsne_result, marker)
}
\arguments{
\item{flowframe}{The flowframe containing the data on which the tSNE was
computed}

\item{tsne_result}{The resulting object of the Rtsne::Rtsne function}

\item{marker}{The marker of interest, given as a character}
}
\value{
ggplot of the tSNE result colored according to the expression of the
provided marker
}
\description{
Function to plot the result of a tSNE, colored according to the expression
values of a certain marker of interest.
}
\examples{


}
