% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_multiple.R
\name{query_multiple}
\alias{query_multiple}
\title{query_multiple}
\usage{
query_multiple(fsom, ff, cell_types, pdf_name, ...)
}
\arguments{
\item{fsom}{FlowSOM object}

\item{ff}{FlowFrame used as reference for channel and marker names}

\item{cell_types}{Description of the cell types. Named list, with one named
vector per cell type containing "high"/"low" values}

\item{pdf_name}{Path to a pdf file to save figures}
}
\value{
A label for every FlowSOM cluster (Unknown or one of the celltype
        names of the list, if selected by QueryStarPlot)
}
\description{
Function which takes a named list of multiple cell types, where every item is
a named vector with values "high"/"low" and the names correspond to the
markers or channels (e.g. as generated by parse_markertable).
}
\examples{


}
